fix(signin): rename description dir
